efield grid corrections for abitrary length
Previous length was coupled to number of grid cells

Also have better understanding of grid cell periodicity that was likely
a bug in previous code. Last "grid point" is really the same one as the
first. So we should wrap around here, I don't think I was accounting for
this and was treating it as a separate point. Fixed now.